claritybot/requirements.txt
streamlit
firebase-admin
google-cloud-firestore
google-cloud-aiplatform
google-adk
pydantic
